+++ missiongen/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.6.10"
+__version__ = "1.7.0"
 
 from .recipe import Recipe
 from .builder import StarterBuilder, generate